issue-111: Engine updated (graphic, physic and multiplayer).
AndEngine updated to anchor center branch.

+++ AndEngine/src/org/andengine/util/adt/list/IIntList.java
@@ -1,7 +1,7 @@
 package org.andengine.util.adt.list;
 
 /**
- * (c) Zynga 2012
+ * (c) 2012 Zynga Inc.
  *
  * @author Nicolas Gramlich <ngramlich@zynga.com>
  * @since 19:21:53 - 03.05.2012
@@ -16,10 +16,10 @@ public interface IIntList {
 	// ===========================================================
 
 	public boolean isEmpty();
-	public float get(final int pIndex) throws ArrayIndexOutOfBoundsException;
+	public int get(final int pIndex) throws ArrayIndexOutOfBoundsException;
 	public void add(final int pItem);
 	public void add(final int pIndex, final int pItem) throws ArrayIndexOutOfBoundsException;
-	public float remove(final int pIndex) throws ArrayIndexOutOfBoundsException;
+	public int remove(final int pIndex) throws ArrayIndexOutOfBoundsException;
 	public int size();
 	public void clear();
 	public int[] toArray();
