+++ AndEngine/src/org/andengine/util/adt/list/ILongList.java
@@ -1,7 +1,7 @@
 package org.andengine.util.adt.list;
 
 /**
- * (c) Zynga 2012
+ * (c) 2012 Zynga Inc.
  *
  * @author Nicolas Gramlich <ngramlich@zynga.com>
  * @since 19:36:57 - 03.05.2012
@@ -16,10 +16,10 @@ public interface ILongList {
 	// ===========================================================
 
 	public boolean isEmpty();
-	public float get(final int pIndex) throws ArrayIndexOutOfBoundsException;
+	public long get(final int pIndex) throws ArrayIndexOutOfBoundsException;
 	public void add(final long pItem);
 	public void add(final int pIndex, final long pItem) throws ArrayIndexOutOfBoundsException;
-	public float remove(final int pIndex) throws ArrayIndexOutOfBoundsException;
+	public long remove(final int pIndex) throws ArrayIndexOutOfBoundsException;
 	public int size();
 	public void clear();
 	public long[] toArray();
